changed end dates of the data products
I have set the end dates of the data products to what they are from the least overlapping time (that was only true for the towers and HErZ, not including ERA20C).
In HandleEraData the minimum of all end dates is determined and from that the data frame filled with all overlapping data. This results in the time series ending with 2010.
What needs to be done now is to figure out how the longest data values can be filled into the data frame while the others are set to NA when not available.